refactor(app): cleanup app and rename root page file to index
